refactor: move search and filter logic out of workout-grid
